pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
    "wheel",
    "setuptools-scm[toml]",
    "numpy>=1.25.0,<2.0.0",
    "mpi4py==3.1.5",
    "swig",
    "Cython",
]
build-backend = "setuptools.build_meta"

[project]
name = "gnrs"
version = "3.0.0"
description = "A random molecular crystal structure generator"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Yi Yang", email = "yiy5@andrew.cmu.edu"},
    {name = "Rithwik Tom"}
]
maintainers = [
    {name = "Yi Yang", email = "yiy5@andrew.cmu.edu"},
]
keywords = ["crystal structure", "molecular crystals", "materials science"]
classifiers = [
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: C"
]

requires-python = ">=3.8,<3.12"
dependencies = [
    "numpy>=1.25.0,<2.0.0",
    "ase",
    "pymatgen>=2023.10.3",
    "matplotlib",
    "pyyaml",
    "tqdm",
    "packaging>=21.2",
    "pyparsing==2.4.7",
    "numba",
    "spglib",
    "mpi4py==3.1.5",
    "torch>=1.12",
    "mace-torch",
    "dscribe",
    "build",
    "twine",
]

[project.urls]
Homepage = "https://github.com/Yi5817/Genarris"
Repository = "https://github.com/Yi5817/Genarris.git"
"Bug Tracker" = "https://github.com/Yi5817/Genarris/issues"

[project.scripts]
gnrs = "gnrs.genarris_master:main"

[tool.setuptools]
packages = ["gnrs"]
include-package-data = true

[tool.setuptools.package-data]
"gnrs.parser" = ["defaults.json"]
"gnrs.generation.cgenarris" = ["src/**/*"]